Generate m3u8 files correctly.
The m3u8 files generated by livepeer were meant for live streaming,
and it had a sliding window of size 6. So, whenever there were more
than 6 segments in total, only the last 6 segments were written
on the m3u8 file. In this commit, we have added a mechanism to
generate a m3u8 file by parsing the segments and finding their
durations. WARNING: this is a very hacky and unoptimized solution!