Always generate m3u8 playlists in Demux
Earlier, we used the playlists generated by livepeer for videos having
less than 6 segments, but livepeer didn't append a "#EXT-X-ENDLIST"
string at the end of the m3u8 file since it was meant for live streaming
using sliding window (of size 6). This commit generates the m3u8 file
in the server itself, and prevents showing loading at the last second
of a video with <= 6 segments.